Moved Adapter/Adapter.php to Adapter.php
